sessions/02_text_sentence_classification/requirements.txt
google-cloud-bigquery[bqstorage,pandas]
google-auth
polars
pyarrow
pandas-gbq
kiwipiepy